requirements.txt
openai>=2.0
google-genai>=1.0
pydantic>=2.0
pyyaml>=6.0
rich>=13.0
python-dotenv>=1.0
tiktoken>=0.5
httpx>=0.25
tavily-python>=0.5
ddgs>=4.0
akshare>=1.10
fastapi>=0.115
uvicorn>=0.30
pytest>=7.0
